requirements.txt
click==8.1.7

Keep crewai at version 0.30.0 as newer versions might have breaking changes.
crewai==0.30.0
crewai-tools==0.1.6
fastapi==0.110.3
google-generativeai==0.5.4
langchain-google-genai==1.0.3
pydantic==2.7.1
python-dotenv==1.0.1
uvicorn==0.29.0
langchain-community==0.0.34